% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_user.R
\name{add_new_user}
\alias{add_new_user}
\title{Add new user to Vedaly}
\usage{
add_new_user(new_user_email, new_user_first_name, new_user_last_name)
}
\arguments{
\item{new_user_email}{The email address of the user.}

\item{new_user_first_name}{First (given) name of the user.}

\item{new_user_last_name}{Last (family) name of the user.}
}
\value{
Invisibly returns `TRUE` if the request was successful.
}
\description{
Add new user to Vedaly
}
