% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_graph.R
\name{share_graph}
\alias{share_graph}
\title{Sign into Vedaly from R}
\usage{
share_graph(
  graph = NULL,
  public = FALSE,
  project = "My Workspace",
  description = "",
  dims = c(800, 400)
)
}
\arguments{
\item{graph}{Base plot function, ggplot2 or plotly object, or NULL for sharing current plot}

\item{public}{Boolean whether the publicly accessible link should be generated}

\item{project}{Project name where the graph should be shared to. By default is it shared in personal space}

\item{description}{Description of the plot, a figure legend.}

\item{dims}{width and height in px}
}
\value{
UUID of the graph
}
\description{
Sign into Vedaly from R
}
