% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_rmd_report.R
\name{share_rmd_report}
\alias{share_rmd_report}
\title{UNDER DEVELOPMENT Share HTML and PDF outputs of rmd to Vedaly}
\usage{
share_rmd_report(rmd_file, project = "My Workspace", title = NULL)
}
\arguments{
\item{rmd_file}{path to .rmd rmarkdown document}

\item{project}{Project name where the report should be shared to. By default is it shared in active project (if set) or in personal workspace}

\item{title}{Report title. Used to display report in the list of reports. By default title from the header is used}
}
\value{
UUID of the report
}
\description{
UNDER DEVELOPMENT Share HTML and PDF outputs of rmd to Vedaly
}
