% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in.R
\name{sign_in}
\alias{sign_in}
\title{Sign into Vedaly from R}
\usage{
sign_in(email)
}
\arguments{
\item{email}{User email}
}
\value{
Invisibly returns TRUE if login was successful.
}
\description{
Sign into Vedaly from R
}
